refactors app into stateless services
