+++ lib/debug-logger.js
@@ -1,6 +1,6 @@
 'use strict'
 
-module.exports = function (controller, config) {
+module.exports.addLogger = (controller, config) => {
   const hearsOrig = controller.hears.bind(controller)
   const formatter = config.formatter || defaultFormatter
